Move email checker to Behat-related services
