+++ src/Sylius/Behat/Resources/config/services/contexts/hook.xml
@@ -24,7 +24,7 @@
         </service>
 
         <service id="sylius.behat.context.hook.email_spool" class="Sylius\Behat\Context\Hook\EmailSpoolContext">
-            <argument type="service" id="__symfony__.sylius.behat.email_checker"/>
+            <argument type="service" id="sylius.behat.email_checker"/>
             <argument type="service" id="__symfony__.filesystem" />
             <tag name="fob.context_service" />
         </service>
