+++ src/Sylius/Bundle/CoreBundle/Resources/config/test_services.xml
@@ -11,17 +11,9 @@
 
 -->
 
-<container xmlns="http://symfony.com/schema/dic/services"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services
-                               http://symfony.com/schema/dic/services/services-1.0.xsd">
+<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
         <!--Paypal-->
         <service id="sylius.invoice_number_generator" class="Sylius\Component\Core\Payment\RandomInvoiceNumberGenerator" />
-
-        <service id="sylius.behat.email_checker" class="Sylius\Component\Core\Test\Services\EmailChecker">
-            <argument>%kernel.cache_dir%/spool</argument>
-        </service>
     </services>
-
 </container>
